refactor(cli): Modularize each function and remove unused code
